requirements.txt
# requirements.txt
streamlit
python-doctr
opencv-python
numpy
Pillow
python-docx
fpdf